Add Raft(Virtual Thread).

+++ src/main/java/io/github/hligaty/raft/config/Configuration.java
@@ -55,7 +55,7 @@ public class Configuration {
         return this;
     }
 
-    public long quorum() {
+    public int quorum() {
         return peers.size() / 2 + 1;
     }
